--- css/index.css
@@ -1,8 +1,0 @@
-
-.red {
-  background: red;
-}
-
-.green {
-  background: green;
-}
