--- markup/components/profile/profile.css
@@ -1,3 +1,0 @@
-.center {
-  text-align: center;
-}
