+++ markup/pages/500/template.js
@@ -1,0 +1,9 @@
+import Handlebars from 'handlebars';
+
+const markup = `
+  <div>
+    Page 500
+  </div>
+`;
+
+export const template = Handlebars.compile(markup);
