+++ markup/pages/chats/template.js
@@ -1,0 +1,9 @@
+import Handlebars from 'handlebars';
+
+const markup = `
+  <div>
+    Chats page
+  </div>
+`;
+
+export const template = Handlebars.compile(markup);
