+++ markup/pages/signup/template.js
@@ -1,0 +1,9 @@
+import Handlebars from 'handlebars';
+
+const markup = `
+  <div>
+    SignUp page
+  </div>
+`;
+
+export const template = Handlebars.compile(markup);
