+++ markup/partials/meta.hbs
@@ -1,20 +1,20 @@
 {{!-- Set character encoding for the document --}}
-<meta charset="utf-8">
+<meta charset='utf-8'>
 {{!-- Instruct Internet Explorer to use its latest rendering engine --}}
-<meta http-equiv="x-ua-compatible" content="ie=edge">
+<meta http-equiv='x-ua-compatible' content='ie=edge'>
 {{!-- Viewport for responsive web design --}}
-<meta name="viewport" content="width=device-width, initial-scale=1">
+<meta name='viewport' content='width=device-width, initial-scale=1'>
 {{#if meta}}
 {{#if meta.author}}
-<meta name="author" content="{{meta.author}}">
+<meta name='author' content='{{meta.author}}'>
 {{/if}}
 {{#if meta.description}}
-<meta name="description" content="{{meta.description}}">
+<meta name='description' content='{{meta.description}}'>
 {{/if}}
 {{#if meta.keywords}}
-<meta name="keywords" content="{{meta.keywords}}">
+<meta name='keywords' content='{{meta.keywords}}'>
 {{/if}}
 {{#if meta.robots}}
-<meta name="robots" content="{{meta.robots}}">
+<meta name='robots' content='{{meta.robots}}'>
 {{/if}}
 {{/if}}
