+++ markup/partials/relationships.hbs
@@ -1,2 +1,2 @@
 {{!-- Recommended favicon format is png --}}
-<link rel="shortcut icon" href="../images/parcel.png" type="image/png">
+<link rel='shortcut icon' href='../images/parcel.png' type='image/png'>
