--- markup/plgd.hbs
@@ -1,21 +1,0 @@
----
-bodyclass: plgd
-language: en
-meta:
-  author: Kirill Efimov
-  description: Playground
-scripts:
-  - ../scripts/plgd.ts
-stylesheets:
-  - ../css/global.css
-title: Playground
-uri: /
----
-
-{{#extend "base"}}
-  {{#content "main"}}
-    <main>
-      <div class='app' />
-    </main>
-  {{/content}}
-{{/extend}}
