--- markup/profile.hbs
@@ -1,23 +1,0 @@
----
-bodyclass: index
-language: en
-meta:
-  author: Kirill Efimov
-  description: Profile
-scripts:
-  - ../scripts/profile.ts
-stylesheets:
-  - ../css/global.css
-  - ../markup/components/profile/profile.css
-title: Profile
-uri: /
----
-
-{{#extend "base"}}
-  {{#content "main"}}
-    <main>
-      <div class='app'>
-      </div>
-    </main>
-  {{/content}}
-{{/extend}}
