--- markup/signup.hbs
@@ -1,76 +1,0 @@
----
-bodyclass: signup
-language: en
-meta:
-  author: Kirill Efimov
-  description: Sugnup page
-stylesheets:
-  - ../css/global.css
-  - ../css/signup.css
-title: Sugnup
-uri: /sugnup
----
-
-{{#extend "base"}}
-  {{#content "main"}}
-    <main>
-      <div class='form-container'>
-        <div class='row center'>
-          <h3>Sign up</h3>
-        </div>
-        <div class='row flex'>
-          <div class='row'>
-            <div>First name</div>
-            <div>
-              <input type='text' class='req-prop'/>
-            </div>
-          </div>    
-          <div class='row ml-2'>
-            <div>Second name</div>
-            <div>
-              <input type='text' class='req-prop'/>
-            </div>
-          </div>    
-        </div>
-        <div class='row'>
-          <div>Email</div>
-          <div>
-            <input type='text' class='req-prop'/>
-          </div>
-        </div>
-        <div class='row'>
-          <div>Login</div>
-          <div>
-            <input type='text' class='req-prop'/>
-          </div>
-        </div>
-        <div class='row'>
-          <div>Phone</div>
-          <div>
-            <input type='text' class='req-prop'/>
-          </div>
-        </div>
-        <div class='row'>
-          <div>Password</div>
-          <div>
-            <input type='text' class='req-prop'/>
-          </div>
-        </div>
-        <div class='row'>
-          <div>Repeat password</div>
-          <div>
-            <input type='text' class='req-prop'/>
-          </div>
-        </div>
-        <div class='row center'>
-          <button id='signin-btn'>
-            <a href="./profile.hbs">Sign up</a>
-          </button>
-        </div>
-        <div class='row center'>
-          <a href="./index.hbs">Sign in</a>
-        </div>
-      </div>
-    </main>
-  {{/content}}
-{{/extend}}
