--- scripts/profile.ts
@@ -1,6 +1,0 @@
-import { render } from '../modules/block';
-import Profile from '../markup/components/profile';
-
-const profile = new Profile();
-
-render('.app', profile);
